feedback changes

+++ jetbrains-core/src/software/aws/toolkits/jetbrains/core/credentials/ConnectionSettingsMenuBuilder.kt
@@ -13,6 +13,7 @@ import com.intellij.openapi.actionSystem.ToggleAction
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.MessageDialogBuilder
 import software.aws.toolkits.core.credentials.CredentialIdentifier
 import software.aws.toolkits.core.region.AwsRegion
 import software.aws.toolkits.jetbrains.core.credentials.sono.SONO_URL
@@ -251,7 +252,13 @@ class ConnectionSettingsMenuBuilder private constructor() {
                     override fun actionPerformed(e: AnActionEvent) {
                         val settings = identitySelectionSettings as? ActionsIdentitySelectionSettings
                         if (value.startUrl != SONO_URL) {
-                            deleteSsoConnectionCW(value)
+                            val confirmDeletion = MessageDialogBuilder.okCancel(
+                                message("gettingstarted.auth.idc.sign.out.confirmation.title"),
+                                message("gettingstarted.auth.idc.sign.out.confirmation")
+                            ).yesText(message("general.confirm")).ask(settings?.project)
+                            if (confirmDeletion) {
+                                deleteSsoConnectionCW(value)
+                            }
                         }
                         logoutFromSsoConnection(settings?.project, value)
                     }
